Create Q242_Valid_Anagram.java
In this code we fill the numbers of characters in strings, into spesific arrays. Arrays length are 26: number of English letters. Our charFiller method's loop checks every letter of string one by one. Every index of our array is increased by one when exact letter is checked. We calculate the letters index's by substracting 97 from their decimal unicode value (decimal value of "a" is 97). Loop's index starts from 0 = a, and ends on 25 = z. Lastly we check if the arrays are equal or not by using Java's "equals" method.