Create Q1913_Product_Difference.java
Our code runs with 3 methods for the best runtime. In the question for the maximum diffrence, we multiply the largest 2 numbers and smallest 2 numbers. We calculate the maximum value by checking every element of array and returning the largest one. Our code also replaces that max value with 0 in array. It helps us get the second largest value afterwards. Same goes for minimum value, except we need to check if the current element is smaller than one. Its because there can't be any number smaller than one in our given array, except the ones we replaced with 0 when we calculate the max values.